# docs/architecture/ml-api-spec.yaml
openapi: 3.0.3
info:
  title: VayuGuard ML Service API
  description: >
    Machine Learning inference service for VayuGuard air quality forecasting
    and health risk assessment. Provides AQI forecasts, personalized health
    risk scores, and model management endpoints.
  version: 1.0.0
  contact:
    name: VayuGuard AI/ML Team
    email: ml-team@vayuguard.com
    url: https://vayuguard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api-staging.vayuguard.com
    description: Staging environment
  - url: https://api.vayuguard.com
    description: Production environment
...
paths:
...
  /api/model/reload:
    post:
      tags: [Model Management]
      summary: Reload model from storage
      description: >
        Trigger a hot-reload of the specified model from the model storage.
        Useful after a new model version is registered via the training pipeline.
        Requires admin authentication.
      operationId: reloadModel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - model_type
              properties:
                model_type:
                  type: string
                  enum: [lstm, xgboost, prophet, ensemble, all]
                  description: Which model to reload
                model_version:
                  type: string
                  description: >
                    Specific version to load. If omitted, loads the latest
                    registered version.
                force:
                  type: boolean
                  default: false
                  description: Force reload even if same version is loaded
      responses:
        '200':
          description: Model reloaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_type:
                    type: string
                  previous_version:
                    type: string
                  current_version:
                    type: string
                  reload_time_ms:
                    type: integer
                  status:
                    type: string
                    enum: [success, skipped]
        '401':
          description: Unauthorized
        '500':
          description: Model reload failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context